Extract lists controller, EditingModels struct, and refresh helper
Move user / extra-services / package edit handlers and the
seed/set_search_results helpers into:
  - controllers/lists.rs : the on_user_*, on_strlist_*, on_pkg_* setup
  - editing_models.rs    : EditingModels struct with attach() and seed()
  - refresh.rs           : refresh_items, called from every controller
                           after a config mutation
