+++ slint-ui/src/controllers/mod.rs
@@ -1,2 +1,3 @@
 pub mod install;
+pub mod lists;
 pub mod welcome;
